+++ src/renderer/src/components/common/TriStateCheckbox.tsx
@@ -1,6 +1,6 @@
 import { InputHTMLAttributes } from 'react';
 
-type TriState = 'checked' | 'unchecked' | 'indeterminate';
+export type TriState = 'checked' | 'unchecked' | 'indeterminate';
 
 type Props = InputHTMLAttributes<HTMLInputElement> & {
   label?: string;
@@ -13,7 +13,7 @@ export const TriStateCheckbox = ({ label, state, onChange, className = '', ...pr
   return (
     <div className={`flex items-center ${className}`}>
       <div
-        className="relative flex items-center justify-center"
+        className="relative flex items-center justify-center cursor-pointer"
         onClick={(e) => {
           e.stopPropagation();
           onChange();